lab 2 part b update

+++ Lab/Lab 2 prac/mainC.c
@@ -11,15 +11,14 @@
 #include <sys/wait.h>
 
 void student_handler(int signum); 
-void rec_handler(int signum); 
 
 int main(){
 
-    // create a shared memory in the main. 
+    // create a shared memory in the main.
     int shmid; 
     key_t key = IPC_PRIVATE; 
     int flag = SHM_R | SHM_W; 
-    shmid = shmget(key , 1000 , flag | IPC_CREAT); 
+    shmid = shmget(key , 1024 , flag | IPC_CREAT); 
     if(shmid < 0){
         perror("shmid"); 
         exit(1); 
@@ -37,39 +36,44 @@ int main(){
     //creating different processes. 
 
     pid_t rpid = fork(); 
+    if(rpid == -1){
+        perror("fork"); 
+        exit(1); 
+    }
     if(rpid == 0){
         // child process => recruiter
-
         // student process created right after
         pid_t spid = fork(); 
+        if(spid == -1){
+            perror("fork"); 
+            exit(1); 
+        }
         if(spid == 0){
-            // student child created by recruiter.
+            // kill(getpid() , SIGSTOP); 
+            // sleep(1); 
+
             signal(SIGUSR1 , student_handler);
             // halt the child processes.
             pause(); 
+            // kill(getpid() , SIGSTOP); 
             int* shms = shm; 
             int s1 = *shms; 
             shms++; 
             int s2 = *shms; 
             shms++;
             int mul = rand()%2; 
-            int ans = (s1+s2)*mul; 
-            *shms = ans;
-            printf("Send kill to parent \n"); 
-            sleep(1); 
-            kill(getppid() , SIGUSR2); 
-            exit(1);    
-        }else if(spid == -1){
-            perror("Fork"); 
-            exit(1); 
+            int ans = (s1+s2)*mul;
+            exit(ans); 
         }else{
-            signal(SIGUSR2 , rec_handler);
-            int num1 = rand()%100; 
-            int num2 = rand()%100;
+            int num1 = rand()%20; 
+            int num2 = rand()%20;
             int* shmr = shm;  
             *shmr = num1;
             shmr++; 
             *shmr = num2;
+            int ans = num1+num2; 
+            
+            // kill(spid , SIGCONT); 
             printf("Send kill to student \n"); 
             sleep(1);  
             int kp = kill(spid , SIGUSR1);
@@ -77,22 +81,17 @@ int main(){
                 perror("Kill Signal"); 
                 exit(1); 
             }
-            pause(); 
-            shmr = shm; 
-            int realAns = num1+num2; 
-            int n1 = *shmr; 
-            shmr++; 
-            int n2 = *shmr; 
-            shmr++; 
-            int temp = *shmr; 
-            printf("%d %d %d" , n1 , n2 ,temp); 
-            if(temp == realAns){
-                printf(" Not an Imposter \n"); 
+            printf("%d %d " , num1 , num2); 
+            int status , retVal; 
+            retVal = waitpid(spid , &status , 0); 
+            printf("%d" , status/256); 
+            if(ans == (status/256)){
+                printf(" Not an imposeter \n"); 
             }else{
-                printf(" Imposter \n"); 
+                printf(" Imposeter \n"); 
             }
         }
-        while(wait(NULL) != -1); 
+        
     }
     while(wait(NULL) != -1); 
     shmdt((void*) shm); 
@@ -102,8 +101,4 @@ int main(){
 
 void student_handler(int signum){
     printf("Send Signal to student \n"); 
-}
-
-void rec_handler(int signum){
-    printf("Send signal to recruiter \n"); 
 }
